パック削除時に製作者IDとログインIDが一致してるか確認。

+++ src/main/java/example/presentation/controller/pack/PackDeleteController.java
@@ -1,6 +1,7 @@
 package example.presentation.controller.pack;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,7 +10,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import example.application.service.PackService;
 import example.domain.model.Session;
+import example.domain.model.UserAuthDetails;
 import example.domain.model.pack.PackId;
+import example.domain.model.user.UserId;
 
 @Controller
 @RequestMapping("packs/{packId}/delete")
@@ -23,17 +26,22 @@ public class PackDeleteController {
     
     @GetMapping
     public String deleteConfirm(Model model,@PathVariable(value = "packId") long packId){
+        
         model.addAttribute("packId", packId);
         return "packs/delete/confirm";
     }
 
     @GetMapping("complete")
-    public String deleteThenRedirect(Model model,@PathVariable(value = "packId") long packId){
-        packService.deletePack(new PackId(packId));
-        if(session.getShuffleList().getSelectIds().contains(packId)){
-            session.getShuffleList().selectIds.remove(session.getShuffleList().selectIds.indexOf(packId));
-        };
-        
-        return "redirect:/top";
+    public String deleteThenRedirect(Model model,@PathVariable(value = "packId") long packId,@AuthenticationPrincipal UserAuthDetails userAuthDetails){
+        UserId userId = packService.getPack(new PackId(packId)).getUserId();
+        if(userAuthDetails.getUserId() == userId){
+            packService.deletePack(new PackId(packId));
+            if(session.getShuffleList().getSelectIds().contains(packId)){
+                session.getShuffleList().selectIds.remove(session.getShuffleList().selectIds.indexOf(packId));
+            };
+            return "redirect:/top";
+        }else{
+            return "redirect:/top";
+        }
     }
 }
